pyproject.toml
[project]
name = "huashu-showcase"
version = "0.1.0"
description = "Huashu Cup optimization and visualization showcase"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "matplotlib>=3.8",
  "pandas>=2.0",
  "openpyxl>=3.1",
  "pillow>=10.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
